+++ src/accounts/margin/MarginWeb3.ts
@@ -15,6 +15,7 @@ import { Buffer } from "buffer"
 import BaseAccount from "../BaseAccount"
 import State from "../State"
 import Control from "../Control"
+import SpecialOrders from "../SpecialOrders"
 import Num from "../../Num"
 import {
 	arePositionsEqual,
@@ -31,6 +32,8 @@ import {
 	ChangeType,
 	ControlSchema,
 	MarginSchema,
+	SpecialOrdersSchema,
+	SpecialOrderType,
 	OrderChangeType,
 	OrderType,
 	OrderTypeName,
@@ -55,7 +58,11 @@ import {
 } from "../../config"
 import Decimal from "decimal.js"
 import { getMintDecimals } from "@zero_one/lite-serum/lib/market"
-import { OrderInfo, PositionInfo } from "../../types/dataTypes"
+import {
+	OrderInfo,
+	PositionInfo,
+	SpecialOrderInfo,
+} from "../../types/dataTypes"
 import EventEmitter from "eventemitter3"
 import {
 	ChangeEvent,
@@ -79,6 +86,7 @@ export interface MarginClassSchema extends Omit<MarginSchema, "collateral"> {
 export default class MarginWeb3 extends BaseAccount<MarginClassSchema> {
 	positions: PositionInfo[] = []
 	orders: OrderInfo[] = []
+	specialOrders: SpecialOrderInfo[] = []
 	_balances: { [key: string]: Num } = {}
 
 	private get rpc(): any {
@@ -120,6 +128,7 @@ export default class MarginWeb3 extends BaseAccount<MarginClassSchema> {
 		pubkey: PublicKey,
 		data: MarginClassSchema,
 		public readonly control: Control,
+		public specialOrdersAccount: SpecialOrders | null,
 		public state: State,
 		private readonly heimdallKey: PublicKey,
 		public readonly owner?: PublicKey,
@@ -140,17 +149,22 @@ export default class MarginWeb3 extends BaseAccount<MarginClassSchema> {
 		simulate = false,
 	): Promise<MarginWeb3> {
 		const marginOwner = owner || program.provider.publicKey!
-		const [[key, _nonce], heimdallKey] = await Promise.all([
+		const [[key, _nonce], soKey, heimdallKey] = await Promise.all([
 			this.getPda(st, marginOwner, program.programId),
+			SpecialOrders.getPda(st, marginOwner, program.programId),
 			this.getHeimdallPda(program.programId),
 		])
 		const data = await this.fetch(program, key, st, commitment)
-		const control = await Control.load(program, data.control, commitment)
+		const [control, specialOrders] = await Promise.all([
+			Control.load(program, data.control, commitment),
+			SpecialOrders.loadNullable(program, soKey),
+		])
 		const margin = new this(
 			program,
 			key,
 			data,
 			control,
+			specialOrders,
 			st,
 			heimdallKey,
 			marginOwner,
@@ -160,6 +174,7 @@ export default class MarginWeb3 extends BaseAccount<MarginClassSchema> {
 		margin.loadBalances()
 		margin.loadPositions()
 		await margin.loadOrders()
+		await margin.loadSpecialOrders()
 		return margin
 	}
 
@@ -175,6 +190,7 @@ export default class MarginWeb3 extends BaseAccount<MarginClassSchema> {
 		commitment?: Commitment,
 		simulate = false,
 		prefetchedHeimdallKey?: PublicKey,
+		prefetchedSpecialOrdersData?: ProgramAccount<SpecialOrdersSchema>,
 	): Promise<MarginWeb3> {
 		const data = this.transformFetchedData(st, prefetchedMarginData.account)
 		const control = await Control.loadPrefetched(
@@ -184,17 +200,21 @@ export default class MarginWeb3 extends BaseAccount<MarginClassSchema> {
 		)
 		const heimdallKey =
 			prefetchedHeimdallKey ??
-			(
-				await PublicKey.findProgramAddress(
-					[Buffer.from("heimdallv1")],
-					program.programId,
-				)
-			)[0]
+			(await this.getHeimdallPda(program.programId))
+		const specialOrders =
+			prefetchedSpecialOrdersData === undefined
+				? null
+				: await SpecialOrders.loadPrefetched(
+						program,
+						prefetchedSpecialOrdersData.publicKey,
+						prefetchedSpecialOrdersData.account,
+				  )
 		const margin = new this(
 			program,
 			prefetchedMarginData.publicKey,
 			data,
 			control,
+			specialOrders,
 			st,
 			heimdallKey,
 			undefined,
@@ -205,6 +225,7 @@ export default class MarginWeb3 extends BaseAccount<MarginClassSchema> {
 		margin.loadPositions()
 		if (withOrders) {
 			await margin.loadOrders()
+			await margin.loadSpecialOrders()
 		}
 		return margin
 	}
@@ -222,20 +243,20 @@ export default class MarginWeb3 extends BaseAccount<MarginClassSchema> {
 			accountInfo.data,
 		)
 		const data = this.transformFetchedData(st, account)
-		const [heimdallKey] = await Promise.all([
-			(
-				await PublicKey.findProgramAddress(
-					[Buffer.from("heimdallv1")],
-					program.programId,
-				)
-			)[0],
+		const [heimdallKey, soKey] = await Promise.all([
+			this.getHeimdallPda(program.programId),
+			await SpecialOrders.getPda(st, data.authority, program.programId),
 		])
-		const control = await Control.load(program, data.control)
+		const [control, specialOrders] = await Promise.all([
+			Control.load(program, data.control),
+			SpecialOrders.loadNullable(program, soKey),
+		])
 		const margin = new this(
 			program,
 			account.publicKey,
 			data,
 			control,
+			specialOrders,
 			st,
 			heimdallKey,
 			undefined,
@@ -246,6 +267,7 @@ export default class MarginWeb3 extends BaseAccount<MarginClassSchema> {
 		margin.loadPositions()
 		if (withOrders) {
 			await margin.loadOrders()
+			await margin.loadSpecialOrders()
 		}
 		return margin
 	}
@@ -263,6 +285,7 @@ export default class MarginWeb3 extends BaseAccount<MarginClassSchema> {
 		this.loadPositions()
 		if (withOrders) {
 			await this.loadOrders()
+			await this.loadSpecialOrders()
 		}
 	}
 
@@ -275,6 +298,7 @@ export default class MarginWeb3 extends BaseAccount<MarginClassSchema> {
 		this.loadPositions()
 		if (withOrders) {
 			await this.loadOrders()
+			await this.loadSpecialOrders()
 		}
 	}
 
@@ -602,6 +626,41 @@ export default class MarginWeb3 extends BaseAccount<MarginClassSchema> {
 		return changeLog
 	}
 
+	async loadSpecialOrders() {
+		if (this.specialOrdersAccount === null) {
+			this.specialOrders = []
+			return
+		}
+
+		const keyToMarket = Object.fromEntries(
+			this.state.data.perpMarkets.map((m) => [m.dexMarket.toString(), m]),
+		)
+
+		this.specialOrders = await Promise.all(
+			this.specialOrdersAccount.data.entries.map(async (x) => {
+				let perpMarket = keyToMarket[x.market.toString()]!
+				let market = await this.state.getMarketBySymbol(
+					perpMarket.symbol,
+				)
+
+				return {
+					id: x.id,
+					marketSymbol: perpMarket.symbol,
+					marketKey: perpMarket.dexMarket,
+					isLong: x.isLong,
+					fee: x.fee,
+					type: x.ty,
+					triggerPrice: new Num(x.triggerPrice, USD_DECIMALS),
+					limitPrice: new Num(x.limitPrice, USD_DECIMALS),
+					size: new Num(
+						market.baseSizeLotsToNumber(x.size),
+						perpMarket.assetDecimals,
+					),
+				}
+			}),
+		)
+	}
+
 	/**
 	 * Refreshes the data on the Margin, state, cache and control accounts.
 	 */
@@ -611,12 +670,20 @@ export default class MarginWeb3 extends BaseAccount<MarginClassSchema> {
 	): Promise<void> {
 		if (refreshMarginData) {
 			if (refreshState) {
-				;[this.data] = await Promise.all([
+				;[this.data, this.specialOrdersAccount] = await Promise.all([
 					MarginWeb3.fetch(
 						this.program,
 						this.pubkey,
 						this.state,
 						this.commitment,
+					),
+					SpecialOrders.loadNullable(
+						this.program,
+						await SpecialOrders.getPda(
+							this.state,
+							this.data.authority,
+							this.program.programId,
+						),
 					),
 					this.control.refresh(),
 					this.state.refresh(),
@@ -633,6 +700,7 @@ export default class MarginWeb3 extends BaseAccount<MarginClassSchema> {
 		this.loadBalances()
 		this.loadPositions()
 		await this.loadOrders()
+		await this.loadSpecialOrders()
 	}
 
 	eventEmitter: EventEmitter<UpdateEvents, ChangeEvent<any>> | null = null
@@ -1660,6 +1728,94 @@ export default class MarginWeb3 extends BaseAccount<MarginClassSchema> {
 				)
 					? ZO_DEX_DEVNET_PROGRAM_ID
 					: ZO_DEX_MAINNET_PROGRAM_ID,
+			},
+		})
+	}
+
+	private async createSpecialOrdersAccountIx(): Promise<
+		[TransactionInstruction, PublicKey]
+	> {
+		const soKey = await SpecialOrders.getPda(
+			this.state,
+			this.data.authority,
+			this.program.programId,
+		)
+		const ix = this.program.instruction.createSpecialOrdersAccount({
+			accounts: {
+				state: this.state.pubkey,
+				authority: this.data.authority,
+				payer: this.data.authority,
+				specialOrders: soKey,
+				rent: SYSVAR_RENT_PUBKEY,
+				systemProgram: SystemProgram.programId,
+			},
+		})
+		return [ix, soKey]
+	}
+
+	async placeSpecialOrder({
+		symbol,
+		isLong,
+		specialOrderType,
+		size,
+		triggerPrice,
+		limitPrice,
+	}: Readonly<{
+		symbol: string
+		isLong: boolean
+		specialOrderType: SpecialOrderType
+		size: number
+		triggerPrice: number
+		limitPrice?: number
+	}>) {
+		const market = await this.state.getMarketBySymbol(symbol)
+		const baseDecimals = this.state.markets[symbol]!.assetDecimals
+		const dexMarket = this.state.getMarketKeyBySymbol(symbol)
+		let [createSoIx, soKey] =
+			this.specialOrdersAccount !== null
+				? [undefined, this.specialOrdersAccount.pubkey]
+				: await this.createSpecialOrdersAccountIx()
+
+		// big / big -> small / big
+		const orderTriggerPrice = new BN(Math.round(triggerPrice * 1e6))
+		const orderLimitPrice =
+			limitPrice !== undefined
+				? new BN(Math.round(limitPrice * 1e6))
+				: isLong
+				? new BN(1)
+				: new BN(1).shln(50)
+		const orderSize = market.baseSizeNumberToLots(size)
+
+		return await this.program.rpc.placeSpecialOrder(
+			isLong,
+			specialOrderType,
+			orderTriggerPrice,
+			orderLimitPrice,
+			orderSize,
+			{
+				accounts: {
+					state: this.state.pubkey,
+					authority: this.data.authority,
+					specialOrders: soKey,
+					dexMarket,
+					systemProgram: SystemProgram.programId,
+				},
+				preInstructions: createSoIx && [createSoIx],
+			},
+		)
+	}
+
+	async cancelSpecialOrder(id: number) {
+		let o = this.specialOrders.find((o) => o.id == id)
+		if (o === undefined) {
+			throw new Error(`Special order ${id} not found`)
+		}
+		return await this.program.rpc.cancelSpecialOrder(id, {
+			accounts: {
+				state: this.state.pubkey,
+				authority: this.data.authority,
+				specialOrders: this.specialOrdersAccount!.pubkey,
+				dexMarket: o.marketKey,
 			},
 		})
 	}